test: refactor and langevin_dynamics initialization test
